Implemented branch prediction optimization
The most common branch path for a materialization is that the object already exists and is the same materialization. Normally this would require 2 queries: 1 information schema query to get the type followed by 1 show query. This runs only the show query in the most common scenario and only runs 2 queries if the object does not exist or exists with a different type. This especially improves the built time for models which are fast to build, such as views, streams, functions, etc.